III-3347 Update or remove dependencies that depend on graceful-fs@2 and @3
For node 10+ support, graceful-fs has to be updated to 4+. So all our dependencies using graceful-fs have to be updated to a version that uses @4.

grunt-google-cdn is also affected but has not been updated in 6 years. See https://github.com/btford/grunt-google-cdn/issues/71
Since it's not a must-have, I've removed it.